First working version of netcode passing things back and forth
