+++ index.html
@@ -1,35 +1,30 @@
 <!DOCTYPE html>
 <html>
-	<head>
+<head>
+    <title>Open Board Game Framework</title>
+    <link rel="stylesheet" href='style.css'>
 
-		<title> Real time multi-player games with HTML5</title>
+    <!-- Library code -->
+        <!-- npm libraries -->
+        <script src="/socket.io/socket.io.js"></script>
+        <!-- local libraries -->
+        <script type="text/javascript" src="lib/dat.gui.min.js"></script>
+        <script type="text/javascript" src="lib/keyboard.js"></script>
 
-			<!-- Load style sheet -->
-		<link href="index.css" media="screen" rel="stylesheet" type="text/css">
+        <!-- our code -->
+        <script type="text/javascript" src="src/game.core.js"></script>
+        <script type="text/javascript" src="client.js"></script>        
 
-	<!-- Library code -->
-		
-		 	<!-- Load socket.io client side contents. -->
-		<script src="/socket.io/socket.io.js"></script>
+</head>
 
-
-	 		<!-- Load the GUI rendering library code. -->
-		<script type="text/javascript" src="lib/dat.gui.min.js"></script>
-			<!-- Load the keyboard specific code --> 
-		<script type="text/javascript" src="lib/keyboard.js"></script>
-	 	
-	<!-- Game code -->
-
-	 		<!-- Load the common game code. -->
-		<script type="text/javascript" src="game.core.js"></script>
-			<!-- Finally, Load the game client code. -->
-		<script type="text/javascript" src="client.js"></script>		
-
-	</head>
-
-	<body>
-		<canvas id="viewport">
-		</canvas>
-	</body>
+<body>
+    <canvas id='game-canvas'>
+    </canvas>
+    <script src='src/main.js'> </script>
+    <script src='src/UI.js'> </script>
+    <script src='src/graphics.js'> </script>
+</body>
+<div id='context-menu'>
+</div>
 
 </html>
