--- old_index.html
@@ -1,25 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <title>Open Board Game Framework</title>
-    <link rel="stylesheet" href='style.css'>
-        <!-- Library code -->
-            <!-- Load socket.io client side contents. -->
-            <script src="/socket.io/socket.io.js"></script>
-            <!-- Load the common game code. -->
-            <script type="text/javascript" src="game.core.js"></script>
-            <!-- Finally, Load the game client code. -->
-            <script type="text/javascript" src="client.js"></script>                
-</head>
-
-<body>
-    <canvas id='game-canvas'>
-    </canvas>
-    <script src='src/main.js'> </script>
-    <script src='src/UI.js'> </script>
-    <script src='src/graphics.js'> </script>
-</body>
-<div id='context-menu'>
-</div>
-
-</html>
